@@ -9,6 +9,7 @@ typedef struct _interval{
 	int a;
 	int b;
 	int begin;
+    int key;
 	GPtrArray *sigma;
 } interval;
 
@@ -19,9 +20,16 @@ typedef struct _interval_pair {
 	int right;
 } interval_pair;
 
-void interval_init(interval *p, int a, int b, GPtrArray* jobarray, int njobs);
+typedef struct _job_interval_pair {
+    Job *j;
+    interval *I;
+    int take;
+} job_interval_pair;
+
+void interval_init(interval *p, int a, int b,int key, GPtrArray* jobarray, int njobs);
+interval *interval_alloc(int a, int b, int key, GPtrArray *jobarray, int njobs);
+interval *interval_copy(interval *src);
 void interval_free(interval *p);
-interval *interval_alloc(int a, int b, GPtrArray *jobarray, int njobs);
 void g_interval_free(void *p);
 
 gint compare_interval(gconstpointer a, gconstpointer b, gpointer data);
