@@ -26,7 +26,15 @@ typedef enum {
     finished = 5,
 } data_status;
 
+/**
+ * problem data
+ */
+typedef struct wctproblem wctproblem;
+/**
+ * node data
+ */
 typedef struct wctdata wctdata;
+
 struct wctdata {
     // The id and depth of the node in the B&B tree
     int id;
@@ -34,34 +42,37 @@ struct wctdata {
 
     data_status status;
 
-    // The job information
+    // The instance information
     int njobs;
     int nmachines;
-    // int *duetime;
-    // int *releasetime;
-    // int *duration;
-    // int *weights;
     int *orig_node_ids;
     // data for meta heuristic
-    Job *jobarray;
-    GPtrArray *jobarray2;
+    GPtrArray *jobarray;
     int  H_max;
     int  H_min;
+    /** data about the intervals */
+    GPtrArray *local_intervals;
+    GPtrArray *ordered_jobs;
+    int **sump;
 
     // The column generation lp information
-    wctlp * LP;
+    wctlp *LP;
+    wctlp *MIP;
     double *x;
     double *coef;
     double *pi;
     // PricerSolver
     PricerSolver *solver;
-    // Colorset(Assignments)
-    int          ccount;
-    scheduleset *cclasses;
+
+    // Columns
+    // int          ccount;
+    // scheduleset *cclasses;
     int          dzcount;
-    int          gallocated;
+    // int          gallocated;
     scheduleset *newsets;
     int          nnewsets;
+    int *cstat;
+    GPtrArray *localColPool;
 
     int  kpc_pi_scalef;
     int  kpc_pi_scalef_heur;
@@ -81,6 +92,7 @@ struct wctdata {
     double  LP_lower_bound_BB;
     double *rhs;
     int     nnonimprovements;
+    int iterations;
     /** Wentges smoothing technique */
     double *pi_in;
     double *pi_out;
@@ -90,6 +102,7 @@ struct wctdata {
     double  eta_in;
     double  eta_out;
     double  eta_sep;
+    double reduced_cost;
     double  alpha;
     int     update;
 
@@ -109,15 +122,15 @@ struct wctdata {
     /** Branching strategies */
     int choose;
     /** conflict */
-    int *    elist_same;
+    int     *elist_same;
     int      ecount_same;
-    int *    elist_differ;
+    int     *elist_differ;
     int      ecount_differ;
     wctdata *same_children;
     int      nsame;
     wctdata *diff_children;
     int      ndiff;
-    int      v1, v2;
+    Job      *v1, *v2;
     /** ahv branching */
     wctdata *duetime_child;
     int      nduetime;
@@ -126,13 +139,22 @@ struct wctdata {
     int      branch_job;
     int      completiontime;
     /** wide branching conflict */
-    int *     v1_wide;
-    int *     v2_wide;
+    int      *v1_wide;
+    int      *v2_wide;
     int       nb_wide;
     wctdata **same_children_wide;
     wctdata **diff_children_wide;
 
+
+    /**
+     * ptr to the parent node
+     */
     wctdata *parent;
+
+    /**
+     * ptr to the data overview
+     */
+    wctproblem *problem;
 
     char pname[MAX_PNAME_LEN];
 };
@@ -149,7 +171,6 @@ typedef enum {
     optimal = 4
 } problem_status;
 
-typedef struct wctproblem wctproblem;
 
 struct wctproblem {
     wctparms parms;
@@ -166,49 +187,65 @@ struct wctproblem {
     int H_min;
     int H_max;
     int off;
-    GPtrArray *e;
     /** nmachines */
     int nmachines;
 
     int    nwctdata;
     int    global_upper_bound;
-    int    first_upper_bound;
     int    global_lower_bound;
-    int    first_lower_bound;
-    double first_rel_error;
     double rel_error;
+    int    root_upper_bound;
+    int    root_lower_bound;
+    double root_rel_error;
     int    maxdepth;
 
     problem_status status;
 
-    /* All stable sets*/
+    /* All partial schedules*/
     scheduleset *initsets;
     int          nbinitsets;
     int          gallocated;
+    GPtrArray *ColPool;
+    /** Maximum number of artificial columns */
+    int maxArtificials;
+    /** Actual number of artificial columns */
+    int nArtificials;
     /* Best Solution*/
-    solution *   opt_sol;
+    solution    *opt_sol;
     scheduleset *bestschedule;
     int          nbestschedule;
     /*heap variables*/
     BinomialHeap *br_heap_a;
-    GPtrArray *   unexplored_states;
-    GQueue *      non_empty_level_pqs;
+    GPtrArray    *unexplored_states;
+    GQueue       *non_empty_level_pqs;
     unsigned int  last_explored;
     int           mult_key;
     int           found;
     int           nb_explored_nodes;
+    int           nb_generated_nodes;
     int           nb_generated_col;
     int           nb_generated_col_root;
     /*Cpu time measurement*/
     CCutil_timer tot_build_dd;
     CCutil_timer tot_cputime;
     CCutil_timer tot_branch_and_bound;
+    CCutil_timer tot_strong_branching;
+    CCutil_timer tot_lb_root;
     CCutil_timer tot_lb;
-    CCutil_timer tot_lb_lp_root;
-    CCutil_timer tot_lb_lp;
+    CCutil_timer tot_solve_lp;
     CCutil_timer tot_pricing;
-    CCutil_timer tot_scatter_search;
-    double       real_time;
+    CCutil_timer tot_heuristic;
+
+    double real_time_build_dd;
+    double real_time_total;
+    double real_time_branch_and_bound;
+    double real_time_strong_branching;
+    double real_time_lb_root;
+    double real_time_lb;
+    double real_time_solve_lp;
+    double real_time_pricing;
+    double real_time_heuristic;
+    //double       real_time;
 };
 
 /*Initialization and free memory for the problem*/
@@ -216,7 +253,7 @@ void wctproblem_init(wctproblem *problem);
 void wctproblem_free(wctproblem *problem);
 
 /*Initialize pmc data*/
-void wctdata_init(wctdata *pd);
+void wctdata_init(wctdata *pd, wctproblem *prob);
 int set_id_and_name(wctdata *pd, int id, const char *fname);
 int wctdata_init_unique(wctdata *pd, int id, const char *name);
 int lp_build(wctdata *pd);
@@ -226,6 +263,26 @@ void lpwctdata_free(wctdata *pd);
 void children_data_free(wctdata *pd);
 void wctdata_free(wctdata *pd);
 void temporary_data_free(wctdata *pd);
+
+/**
+ * solver zdd
+ */
+int solve_weight_dbl_bdd(wctdata *pd);
+int solve_weight_dbl_zdd(wctdata *pd);
+int solve_pricing(wctdata *pd, wctparms *parms, int evaluate);
+int solve_farkas_dbl(wctdata *pd);
+int solve_farkas_dbl_DP(wctdata *pd);
+void print_dot_file(PricerSolver *solver, char *name);
+int evaluate_nodes(wctdata *pd);
+int calculate_new_ordered_jobs(wctdata *pd);
+int build_solve_mip(wctdata *pd);
+void print_number_nodes_edges(wctdata *pd);
+
+/**
+ * Stabilization techniques
+ */
+int solve_stab(wctdata *pd, wctparms *parms);
+int solve_stab_dynamic(wctdata *pd, wctparms *parms);
 #ifdef __cplusplus
 }
 #endif
