@@ -1,25 +1,23 @@
 #include <interval.h>
 
-void g_print_interval(gpointer data, gpointer user_data)
-{
+void g_print_interval(gpointer data, gpointer user_data) {
     interval *a = (interval *)data;
-    printf("%d %d: ", a->a, a->b);
+    printf("interval %d: (%d %d]: ", a->key, a->a, a->b);
     g_ptr_array_foreach(a->sigma, g_print_job, NULL);
     printf("\n");
 }
 
-gint compare_interval(gconstpointer a, gconstpointer b, gpointer data)
-{
-    const Job *x = *(Job * const *)a;
-    const Job *y = *(Job * const *)b;
-    interval *user_data = (interval *)data;
+gint compare_interval(gconstpointer a, gconstpointer b, gpointer data) {
+    const Job *x = *(Job *const *)a;
+    const Job *y = *(Job *const *)b;
+    interval * user_data = (interval *)data;
     int        diff = user_data->b - user_data->a;
-    double w_x = (x->duetime <= user_data->a)
-                 ? (double) x->weight / x->processingime
-                 : 0.0;
+    double     w_x = (x->duetime <= user_data->a)
+                     ? (double)x->weight / x->processingime
+                     : 0.0;
     double w_y = (y->duetime <= user_data->a)
-                 ? (double) y->weight / y->processingime
-                 : 0.0;
+                     ? (double)y->weight / y->processingime
+                     : 0.0;
 
     if (x->processingime > diff) {
         if (y->processingime <= diff) {
@@ -35,7 +33,7 @@ gint compare_interval(gconstpointer a, gconstpointer b, gpointer data)
                 return 1;
             }
 
-            return 0;
+            return x->job - y->job;
         }
     } else {
         if (y->processingime > diff) {
@@ -51,15 +49,16 @@ gint compare_interval(gconstpointer a, gconstpointer b, gpointer data)
                 return 1;
             }
 
-            return 0;
+            return x->job - y->job;
         }
     }
 }
 
-void interval_init(interval *p, int a, int b, GPtrArray *jobarray, int njobs)
-{
+void interval_init(
+    interval *p, int a, int b, int key, GPtrArray *jobarray, int njobs) {
     p->a = a;
     p->b = b;
+    p->key = key;
     p->sigma = g_ptr_array_new();
     p->begin = 0;
     Job *j;
@@ -70,31 +69,46 @@ void interval_init(interval *p, int a, int b, GPtrArray *jobarray, int njobs)
 
     g_ptr_array_sort_with_data(p->sigma, compare_interval, p);
 
-    j = (Job *) g_ptr_array_index(p->sigma, p->begin);
-    while(p->b - p->a <= j->processingime && p->begin < (int)jobarray->len) {
-        j = (Job *) g_ptr_array_index(p->sigma,p->begin);
+    j = (Job *)g_ptr_array_index(p->sigma, p->begin);
+    while (p->b - p->a <= j->processingime && p->begin < (int)jobarray->len) {
+        j = (Job *)g_ptr_array_index(p->sigma, p->begin);
         p->begin++;
     }
 }
 
-interval *interval_alloc(int a, int b, GPtrArray *jobarray, int njobs)
-{
+interval *interval_alloc(
+    int a, int b, int key, GPtrArray *jobarray, int njobs) {
     interval *p = CC_SAFE_MALLOC(1, interval);
     CCcheck_NULL_3(p, "Failed to allocate memory")
-    interval_init(p, a, b, jobarray, njobs);
+        interval_init(p, a, b, key, jobarray, njobs);
 CLEAN:
     return p;
 }
 
-void interval_free(interval *p)
-{
+interval *interval_copy(interval *src) {
+    interval *ret = CC_SAFE_MALLOC(1, interval);
+    CCcheck_NULL_3(ret, "Failed to allocate memory");
+
+    ret->a = src->a;
+    ret->b = src->b;
+    ret->begin = src->begin;
+
+    g_ptr_array_sized_new(src->sigma->len);
+
+    for (unsigned i = 0; i < src->sigma->len; ++i) {
+        g_ptr_array_add(ret->sigma, g_ptr_array_index(src->sigma, i));
+    }
+CLEAN:
+    return ret;
+}
+
+void interval_free(interval *p) {
     if (p != (interval *)NULL) {
         g_ptr_array_free(p->sigma, TRUE);
     }
 }
 
-void g_interval_free(void *p)
-{
+void g_interval_free(void *p) {
     if (p != NULL) {
         interval *part = (interval *)p;
         g_ptr_array_free(part->sigma, TRUE);
@@ -102,10 +116,9 @@ void g_interval_free(void *p)
     }
 }
 
-void interval_pair_free(void *p){
-    if(p != NULL) {
-        interval_pair *tmp = (interval_pair *) p;
+void interval_pair_free(void *p) {
+    if (p != NULL) {
+        interval_pair *tmp = (interval_pair *)p;
         CC_FREE(tmp, interval_pair);
     }
 }
-
