@@ -1,44 +1,43 @@
 #include <wct.h>
 
 int grab_int_sol(wctdata *pd, double *x, double tolerance) {
-    int    val = 0;
-    double test_incumbent = .0;
-    double incumbent;
-    int    i;
-    int    tot_weighted = 0;
-    int *  colored = CC_SAFE_MALLOC(pd->njobs, int);
-    CCcheck_NULL_2(colored, "Failed to allocate colored");
-    fill_int(colored, pd->njobs, 0);
+    int          val = 0;
+    double       test_incumbent = .0;
+    double       incumbent;
+    int          i;
+    int          tot_weighted = 0;
+    int          nb_cols;
+    scheduleset *tmp_schedule;
+    Job *        tmp_j;
+
     val = wctlp_objval(pd->LP, &incumbent);
     CCcheck_val_2(val, "wctlp_objval failed");
+    val = wctlp_get_nb_cols(pd->LP, &nb_cols);
+    CCcheck_val_2(val, "Failed get nb_cols");
+
     schedulesets_free(&(pd->bestcolors), &(pd->nbbest));
     pd->bestcolors = CC_SAFE_MALLOC(pd->nmachines, scheduleset);
     CCcheck_NULL_2(pd->bestcolors, "Failed to realloc pd->bestcolors");
     pd->nbbest = 0;
 
-    for (i = 0; i < pd->ccount; ++i) {
+    assert(nb_cols == pd->localColPool->len);
+    for (i = 0; i < pd->localColPool->len; ++i) {
+        tmp_schedule = (scheduleset *)g_ptr_array_index(pd->localColPool, i);
         test_incumbent += x[i];
 
         if (x[i] >= 1.0 - tolerance) {
             int j = pd->nbbest;
             int k;
             scheduleset_init(pd->bestcolors + j);
-            pd->bestcolors[j].members =
-                CC_SAFE_MALLOC(pd->cclasses[i].count, int);
-            CCcheck_NULL_2(pd->bestcolors[j].members,
-                           "Failed to realloc pd->bestcolors[j].members");
 
-            for (k = 0; k < pd->cclasses[i].count; ++k) {
-                if (!colored[pd->cclasses[i].members[k]]) {
-                    colored[pd->cclasses[i].members[k]] = 1;
-                    pd->bestcolors[j].members[pd->bestcolors[j].count++] =
-                        pd->cclasses[i].members[k];
-                    pd->bestcolors[j].totweight +=
-                        pd->jobarray[pd->cclasses[i].members[k]].processingime;
-                    pd->bestcolors[j].totwct +=
-                        pd->jobarray[pd->cclasses[i].members[k]].weight *
-                        pd->bestcolors[j].totweight;
-                }
+            g_ptr_array_set_size(pd->bestcolors[j].jobs,
+                                 tmp_schedule->jobs->len);
+            for (k = 0; k < tmp_schedule->jobs->len; ++k) {
+                tmp_j = (Job *)g_ptr_array_index(tmp_schedule->jobs, k);
+                g_ptr_array_add(pd->bestcolors[j].jobs, tmp_j);
+                pd->bestcolors[j].totweight += tmp_j->processingime;
+                pd->bestcolors[j].totwct +=
+                    value_Fj(pd->bestcolors[j].totweight, tmp_j);
             }
 
             pd->nbbest++;
@@ -55,8 +54,7 @@ int grab_int_sol(wctdata *pd, double *x, double tolerance) {
         }
     }
 
-    val = scheduleset_check(pd->bestcolors, pd->nbbest, pd->njobs);
-    CCcheck_val_2(val, "ERROR: An incorrect coloring was created.");
+    /** Write a check function */
     printf("Intermediate schedule:\n");
     print_schedule(pd->bestcolors, pd->nbbest);
     printf("with total weight %d\n", tot_weighted);
@@ -72,119 +70,156 @@ int grab_int_sol(wctdata *pd, double *x, double tolerance) {
     }
 
 CLEAN:
-    CC_IFFREE(colored, int);
     return val;
 }
 
+int addColToLP(scheduleset *set, wctdata *pd) {
+    int        val = 0;
+    int        cind;
+    int        vind;
+    double     cval;
+    int        njobs = pd->njobs;
+    GPtrArray *members = set->jobs;
+    wctlp *    lp = pd->LP;
+    Job *      job;
+
+    val = wctlp_get_nb_cols(lp, &(set->id));
+    CCcheck_val_2(val, "Failed to get the number of cols");
+    vind = set->id;
+    val = wctlp_addcol(lp, 0, NULL, NULL, (double)set->totwct, 0.0,
+                       GRB_INFINITY, wctlp_CONT, NULL);
+    CCcheck_val_2(val, "Failed to add column to lp")
+
+        for (unsigned i = 0; i < members->len; ++i) {
+        job = (Job *)g_ptr_array_index(members, i);
+        cind = job->job;
+        val = wctlp_getcoef(lp, &cind, &vind, &cval);
+        CCcheck_val_2(val, "Failed wctlp_getcoef");
+        cval += 1.0;
+        set->nb[job->job] += 1;
+        val = wctlp_chgcoef(lp, 1, &cind, &vind, &cval);
+        CCcheck_val_2(val, "Failed wctlp_chgcoef");
+    }
+
+    cind = njobs;
+    cval = 1.0;
+    val = wctlp_chgcoef(lp, 1, &cind, &vind, &cval);
+    CCcheck_val_2(val, "Failed wctlp_chgcoef");
+
+CLEAN:
+    return val;
+}
+
+void g_add_col_to_lp(gpointer data, gpointer user_data) {
+    scheduleset *tmp = (scheduleset *)data;
+    wctdata *    pd = (wctdata *)user_data;
+    addColToLP(tmp, pd);
+}
+
 int build_lp(wctdata *pd, int construct) {
-    int  val = 0;
-    int  i, j;
-    int  counter = 0;
-    int *covered = CC_SAFE_MALLOC(pd->njobs, int);
+    int         val = 0;
+    wctproblem *problem = pd->problem;
+    int         njobs = problem->njobs;
+    int         nmachines = problem->nmachines;
+    int         nb_row;
+    int *       covered = CC_SAFE_MALLOC(njobs, int);
     CCcheck_NULL_2(covered, "Failed to allocate memory to covered");
-    fill_int(covered, pd->njobs, 0);
+    fill_int(covered, njobs, 0);
     val = wctlp_init(&(pd->LP), NULL);
     CCcheck_val_2(val, "wctlp_init failed");
 
-    for (i = 0; i < pd->njobs; i++) {
-        val = wctlp_addrow(pd->LP, 0, (int *)NULL, (double *)NULL,
-                           wctlp_GREATER_EQUAL, 1.0, (char *)NULL);
+    for (int i = 0; i < njobs; i++) {
+        val = wctlp_addrow(pd->LP, 0, (int *)NULL, (double *)NULL, wctlp_GREATER_EQUAL,
+                           1.0, (char *)NULL);
         CCcheck_val_2(val, "Failed wctlp_addrow");
     }
 
-    wctlp_addrow(pd->LP, 0, (int *)NULL, (double *)NULL, wctlp_LESS_EQUAL,
-                 (double)pd->nmachines, NULL);
-    pd->coef =
-        (double *)CCutil_reallocrus(pd->coef, (pd->njobs + 1) * sizeof(double));
-    CCcheck_NULL_2(pd->coef, "out of memory for coef");
-    fill_dbl(pd->coef, pd->njobs + 1, 1.0);
+    val = wctlp_addrow(pd->LP, 0, (int *)NULL, (double *)NULL, wctlp_LESS_EQUAL,
+                       (double)nmachines, (char *)NULL);
+
+    wctlp_get_nb_rows(pd->LP, &nb_row);
+
+    /** add columns from localColPool */
+    g_ptr_array_foreach(pd->localColPool, g_add_col_to_lp, pd);
 
     /** add constraint about number of machines */
-    if (construct) {
-        for (i = 0; i < pd->ccount; i++) {
-            val = wctlp_addcol(pd->LP, pd->cclasses[i].count + 1,
-                               pd->cclasses[i].members, pd->coef,
-                               pd->cclasses[i].totwct, 0.0, GRB_INFINITY,
-                               wctlp_CONT, NULL);
-
-            if (val) {
-                wctlp_printerrorcode(val);
-            }
-
-            for (j = 0; j < pd->cclasses[i].count && counter < pd->njobs; j++) {
-                if (!covered[pd->cclasses[i].members[j]]) {
-                    covered[pd->cclasses[i].members[j]] = 1;
-                    counter++;
-                }
-            }
-        }
-    }
-
-    if (counter < pd->njobs) {
-        /** Farkas Pricing */
-        for (i = 0; i < pd->njobs; i++) {
-            if (!covered[i]) {
-                pd->nnewsets = 1;
-                pd->newsets = CC_SAFE_MALLOC(1, scheduleset);
-                scheduleset_init(pd->newsets);
-                pd->newsets[0].members = CC_SAFE_MALLOC(2, int);
-                CCcheck_NULL_2(
-                    pd->newsets[0].members,
-                    "Failed to allocate memory to pd->newsets->members");
-                pd->newsets[0].C = CC_SAFE_MALLOC(1, int);
-                CCcheck_NULL_2(pd->newsets[0].C, "Failed to allocate memory");
-                pd->newsets[0].table =
-                    g_hash_table_new(g_direct_hash, g_direct_equal);
-                CCcheck_NULL_2(pd->newsets[0].table,
-                               "Failed to allocate memory");
-                pd->newsets[0].count++;
-                pd->newsets[0].members[0] = i;
-                pd->newsets[0].members[1] = pd->njobs;
-                pd->newsets[0].totwct =
-                    pd->jobarray[i].weight * pd->jobarray[i].processingime;
-                pd->newsets[0].totweight = pd->jobarray[i].processingime;
-                pd->newsets[0].C[0] = pd->jobarray[i].processingime;
-                g_hash_table_insert(pd->newsets[0].table, GINT_TO_POINTER(i),
-                                    pd->newsets[0].C);
-                pd->newsets->age = 0;
-                val = wctlp_addcol(pd->LP, 2, pd->newsets[0].members, pd->coef,
-                                   pd->newsets[0].totwct, 0.0, GRB_INFINITY,
-                                   wctlp_CONT, NULL);
-                CCcheck_val_2(val, "Failed in wctlp_addcol");
-
-                if (pd->gallocated == 0 && pd->ccount == 0) {
-                    pd->cclasses = CC_SAFE_MALLOC(pd->njobs, scheduleset);
-
-                    for (j = 0; j < pd->njobs; ++j) {
-                        scheduleset_init(pd->cclasses + i);
-                    }
-
-                    pd->gallocated = pd->njobs;
-                }
-
-                add_newsets(pd);
-            }
-        }
-    }
-
-    pd->pi =
-        (double *)CCutil_reallocrus(pd->pi, (pd->njobs + 1) * sizeof(double));
+    // if (construct) {
+    //     for (i = 0; i < pd->ccount; i++) {
+    //         val = wctlp_addcol(pd->LP, pd->cclasses[i].count + 1,
+    //                            pd->cclasses[i].members, pd->coef,
+    //                            pd->cclasses[i].totwct, 0.0, GRB_INFINITY,
+    //                            wctlp_CONT, NULL);
+    //         if (val) {
+    //             wctlp_printerrorcode(val);
+    //         }
+    //         for (j = 0; j < pd->cclasses[i].count && counter < njobs; j++) {
+    //             if (!covered[pd->cclasses[i].members[j]]) {
+    //                 covered[pd->cclasses[i].members[j]] = 1;
+    //                 counter++;
+    //             }
+    //         }
+    //     }
+    // }
+    // if (counter < njobs) {
+    //     /** Farkas Pricing */
+    //     for (i = 0; i < njobs; i++) {
+    //         if (!covered[i]) {
+    //             pd->nnewsets = 1;
+    //             pd->newsets = CC_SAFE_MALLOC(1, scheduleset);
+    //             scheduleset_init(pd->newsets);
+    //             pd->newsets[0].members = CC_SAFE_MALLOC(2, int);
+    //             CCcheck_NULL_2(
+    //                 pd->newsets[0].members,
+    //                 "Failed to allocate memory to pd->newsets->members");
+    //             pd->newsets[0].C = CC_SAFE_MALLOC(1, int);
+    //             CCcheck_NULL_2(pd->newsets[0].C, "Failed to allocate
+    //             memory"); pd->newsets[0].table =
+    //                 g_hash_table_new(g_direct_hash, g_direct_equal);
+    //             CCcheck_NULL_2(pd->newsets[0].table,
+    //                            "Failed to allocate memory");
+    //             pd->newsets[0].count++;
+    //             pd->newsets[0].members[0] = i;
+    //             pd->newsets[0].members[1] = njobs;
+    //             pd->newsets[0].totwct =
+    //                 pd->jobarray[i].weight * pd->jobarray[i].processingime;
+    //             pd->newsets[0].totweight = pd->jobarray[i].processingime;
+    //             pd->newsets[0].C[0] = pd->jobarray[i].processingime;
+    //             g_hash_table_insert(pd->newsets[0].table, GINT_TO_POINTER(i),
+    //                                 pd->newsets[0].C);
+    //             pd->newsets->age = 0;
+    //             val = wctlp_addcol(pd->LP, 2, pd->newsets[0].members,
+    //             pd->coef,
+    //                                pd->newsets[0].totwct, 0.0, GRB_INFINITY,
+    //                                wctlp_CONT, NULL);
+    //             CCcheck_val_2(val, "Failed in wctlp_addcol");
+    //             if (pd->gallocated == 0 && pd->ccount == 0) {
+    //                 pd->cclasses = CC_SAFE_MALLOC(njobs, scheduleset);
+    //                 for (j = 0; j < njobs; ++j) {
+    //                     scheduleset_init(pd->cclasses + i);
+    //                 }
+    //                 pd->gallocated = njobs;
+    //             }
+    //             add_newsets(pd);
+    //         }
+    //     }
+    // }
+    pd->pi = (double *)CCutil_reallocrus(pd->pi, nb_row * sizeof(double));
     CCcheck_NULL_2(pd->pi, "Failed to allocate memory to pd->pi");
-    pd->pi_in = CC_SAFE_MALLOC(pd->njobs + 1, double);
+    pd->pi_in = CC_SAFE_MALLOC(nb_row, double);
     CCcheck_NULL_2(pd->pi_in, "Failed to allocate memory");
-    fill_dbl(pd->pi_in, pd->njobs + 1, 0.0);
+    fill_dbl(pd->pi_in, nb_row, 0.0);
     pd->eta_in = 0.0;
-    pd->pi_out = CC_SAFE_MALLOC(pd->njobs + 1, double);
+    pd->pi_out = CC_SAFE_MALLOC(nb_row, double);
     CCcheck_NULL_2(pd->pi_out, "Failed to allocate memory");
     pd->eta_out = 0.0;
-    fill_dbl(pd->pi_out, pd->njobs + 1, 0.0);
-    pd->pi_sep = CC_SAFE_MALLOC(pd->njobs + 1, double);
+    fill_dbl(pd->pi_out, nb_row, 0.0);
+    pd->pi_sep = CC_SAFE_MALLOC(nb_row, double);
     CCcheck_NULL_2(pd->pi_sep, "Failed to allocate memory");
-    pd->subgradient_in = CC_SAFE_MALLOC(pd->njobs + 1, double);
+    pd->subgradient_in = CC_SAFE_MALLOC(nb_row, double);
     CCcheck_NULL_2(pd->subgradient_in, "Failed to allocate memory");
-    pd->subgradient = CC_SAFE_MALLOC(pd->njobs + 1, double);
+    pd->subgradient = CC_SAFE_MALLOC(nb_row, double);
     CCcheck_NULL_2(pd->subgradient, "Failed to allocate memory");
-    pd->rhs = CC_SAFE_MALLOC(pd->njobs + 1, double);
+    pd->rhs = CC_SAFE_MALLOC(nb_row, double);
     CCcheck_NULL_2(pd->rhs, "Failed to allocate memory");
     val = wctlp_get_rhs(pd->LP, pd->rhs);
     CCcheck_val_2(val, "Failed to get RHS");
@@ -203,5 +238,36 @@ CLEAN:
     }
 
     CC_IFFREE(covered, int);
+    return val;
+}
+
+int get_solution_lp_lowerbound(wctdata *pd) {
+    int          val = 0;
+    int          nbcols;
+    scheduleset *tmp;
+    Job *        tmp_j;
+
+    val = wctlp_get_nb_cols(pd->LP, &nbcols);
+    CCcheck_val_2(val, "Failed in wctlp_get_nb_cols");
+    pd->x = CC_SAFE_REALLOC(pd->x, nbcols, double);
+    CCcheck_NULL_2(pd->x, "Failed to allocate memory");
+    assert(nbcols == pd->localColPool->len);
+    wctlp_x(pd->LP, pd->x, 0);
+
+    for (unsigned i = 0; i < nbcols; ++i) {
+        tmp = ((scheduleset *)g_ptr_array_index(pd->localColPool, i));
+        if (pd->x[i]) {
+            printf("%f: ", pd->x[i]);
+            g_ptr_array_foreach(tmp->jobs, g_print_machine, NULL);
+            printf("\n");
+            for (unsigned j = 0; j < tmp->jobs->len; ++j) {
+                tmp_j = (Job *)g_ptr_array_index(tmp->jobs, j);
+                printf("%d (%d) ", tmp->nb[tmp_j->job], tmp_j->processingime);
+            }
+            printf("\n");
+        }
+    }
+
+CLEAN:
     return val;
 }
