@@ -13,8 +13,8 @@ void g_problem_summary_init(gpointer data, gpointer user_data) {
 }
 
 gint g_job_compare_edd(const void *a, const void *b, void *data) {
-    const Job *x = *((Job * const *)a);
-    const Job *y = *((Job * const *)b);
+    const Job *x = *((Job *const *)a);
+    const Job *y = *((Job *const *)b);
 
     if (x->duetime > y->duetime) {
         return (1);
@@ -24,9 +24,9 @@ gint g_job_compare_edd(const void *a, const void *b, void *data) {
         return (1);
     } else if (x->processingime < y->processingime) {
         return (-1);
-    } else if (x->weight > y->weight) {
-        return (1);
     } else if (x->weight < y->weight) {
+        return (1);
+    } else if (x->weight > y->weight) {
         return (-1);
     } else if (x->job > y->job) {
         return (1);
@@ -35,22 +35,6 @@ gint g_job_compare_edd(const void *a, const void *b, void *data) {
     }
 
     return (0);
-}
-
-int calculate_Hmax(Job *jobarray, int nmachines, int njobs) {
-    int    i, max = jobarray[0].processingime, val = 0;
-    double temp;
-
-    for (i = 0; i < njobs; ++i) {
-        max = CC_MAX(jobarray[i].processingime, max);
-        val += jobarray[i].processingime;
-    }
-
-    val -= max;
-    temp = (double)val;
-    temp = temp / (double)nmachines + max;
-    val = (int)ceil(temp);
-    return val;
 }
 
 int calculate_Hmin(
@@ -72,37 +56,84 @@ int calculate_Hmin(
     return val;
 }
 
-int preprocess_data(wctproblem *problem) {
-    int      val = 0;
+void calculate_Hmax(wctproblem *problem) {
     int      temp = 0;
     double   temp_dbl = 0.0;
     wctdata *pd = &(problem->root_pd);
 
+    temp = problem->psum - problem->pmax;
+    temp_dbl = (double)temp;
+    temp_dbl = floor(temp_dbl / problem->nmachines);
+    problem->H_max = pd->H_max = (int)temp_dbl + problem->pmax + 10;
+    problem->H_min = (int)ceil(temp_dbl / problem->nmachines) - problem->pmax;
+    printf("H_max = %d,  pmax = %d, pmin = %d, psum = %d, off = %d\n",
+           problem->H_max, problem->pmax, problem->pmin, problem->psum,
+           problem->off);
+}
+
+void determine_jobs_order_interval(wctproblem *problem) {
+    Job *     tmp, *tmp_j;
+    interval *tmp_interval;
+
+    GPtrArray *local_intervals = problem->root_pd.local_intervals;
+
+    for (unsigned i = 0; i < problem->g_job_array->len; ++i) {
+        tmp_j = (Job *)g_ptr_array_index(problem->g_job_array, i);
+        tmp_j->pos_interval = CC_SAFE_MALLOC(local_intervals->len, int);
+        for (unsigned j = 0; j < local_intervals->len; ++j) {
+            tmp_interval = (interval *)g_ptr_array_index(local_intervals, j);
+            GPtrArray *sigma = tmp_interval->sigma;
+            for (unsigned k = 0; k < sigma->len; ++k) {
+                tmp = (Job *)g_ptr_array_index(sigma, k);
+                if (tmp == tmp_j) {
+                    tmp_j->pos_interval[j] = k;
+                    break;
+                }
+            }
+        }
+    }
+}
+
+int preprocess_data(wctproblem *problem) {
+    int val = 0;
+    int i = 0;
+    wctdata *root = &(problem->root_pd);
+
+    /** Calculate the statistics of the instance */
     g_ptr_array_foreach(problem->g_job_array, g_problem_summary_init, problem);
 
     /** Calculate H_max */
-    temp = problem->psum - problem->pmax;
-    temp_dbl = (double)temp;
-    temp_dbl = temp_dbl / problem->nmachines + problem->pmax;
-    problem->H_max = pd->H_max = (int)ceil(temp_dbl);
-    printf("H_max = %d\n", problem->H_max);
+    calculate_Hmax(problem);
 
+    /** order the jobarray of problem following edd rule */
     g_ptr_array_sort_with_data(problem->g_job_array, g_job_compare_edd, NULL);
 
+    g_ptr_array_foreach(problem->g_job_array, g_set_jobarray_job, &i);
+    root->jobarray = problem->g_job_array;
+
+    /** Find the intervals of the instance at the root node */
     find_division(problem);
+
+    /** Create all node of the ZDD */
+    create_ordered_jobs_array(root->local_intervals, root->ordered_jobs);
+
+    /** Determine the position of each job in the interval */
+    determine_jobs_order_interval(problem);
+
+    g_ptr_array_foreach(root->local_intervals, g_print_interval, NULL);
 
     return val;
 }
 
 static int calculate_T(interval_pair *pair, int k, GPtrArray *interval_array) {
     interval *I = (interval *)g_ptr_array_index(interval_array, k);
-    interval * tmp;
-    Job *i = pair->a;
-    Job *j = pair->b;
+    interval *tmp;
+    Job *     i = pair->a;
+    Job *     j = pair->b;
     pair->left = I->a;
     pair->right = I->a + j->processingime;
 
-    if ( pair->left > I->b - i->processingime) {
+    if (pair->left > I->b - i->processingime) {
         return pair->left;
     } else {
         if (value_diff_Fij(pair->left, i, j) <= 0) {
@@ -113,7 +144,7 @@ static int calculate_T(interval_pair *pair, int k, GPtrArray *interval_array) {
             tmp = (interval *)g_ptr_array_index(interval_array, t);
             pair->left = tmp->a + j->processingime - i->processingime;
 
-            if (value_diff_Fij(pair->left, i, j) <= 0) {
+            if (value_diff_Fij(pair->left, i, j) <= 0 && pair->left >= tmp->a && pair->left <= tmp->b - i->processingime) {
                 break;
             }
         }
@@ -122,33 +153,35 @@ static int calculate_T(interval_pair *pair, int k, GPtrArray *interval_array) {
     }
 }
 
-static int check_interval(interval_pair *pair, int k, GPtrArray *interval_array) {
+static int check_interval(interval_pair *pair,
+                          int            k,
+                          GPtrArray *    interval_array) {
     interval *I = (interval *)g_ptr_array_index(interval_array, k);
-    Job *j = pair->b;
+    Job *     j = pair->b;
     return (I->a + j->processingime >= I->b ||
             calculate_T(pair, k, interval_array) <= I->a);
 }
 
-static GPtrArray *array_time_slots(interval *I, GList *pairs){
-    GPtrArray *array = g_ptr_array_new_with_free_func(free);
+static GPtrArray *array_time_slots(interval *I, GList *pairs) {
+    GPtrArray *    array = g_ptr_array_new_with_free_func(free);
     interval_pair *tmp;
     interval_pair *min_data;
-    GList *min;
-    int *tmp_int, prev;
+    GList *        min;
+    int *          tmp_int, prev;
 
     tmp_int = CC_SAFE_MALLOC(1, int);
     *tmp_int = I->a;
     g_ptr_array_add(array, tmp_int);
     prev = *tmp_int;
 
-    while(pairs) {
+    while (pairs) {
         min = pairs;
-        min_data = (interval_pair *) min->data;
-        for(GList * i = min->next; i; i = g_list_next(i)) {
-            tmp = ((interval_pair *) i->data);
-            if(tmp->right < min_data->right) {
+        min_data = (interval_pair *)min->data;
+        for (GList *i = min->next; i; i = g_list_next(i)) {
+            tmp = ((interval_pair *)i->data);
+            if (tmp->right < min_data->right) {
                 min = i;
-                min_data = (interval_pair *) i->data;
+                min_data = (interval_pair *)i->data;
             }
         }
 
@@ -159,15 +192,14 @@ static GPtrArray *array_time_slots(interval *I, GList *pairs){
         g_list_free_full(min, interval_pair_free);
 
         GList *i = pairs;
-        while(i){
-            tmp = (interval_pair *) i->data;
-            if(*tmp_int >= tmp->left && *tmp_int <= tmp->right) {
+        while (i) {
+            tmp = (interval_pair *)i->data;
+            if (*tmp_int >= tmp->left && *tmp_int <= tmp->right) {
                 GList *remove = i;
                 i = g_list_next(i);
                 pairs = g_list_remove_link(pairs, remove);
                 g_list_free_full(remove, interval_pair_free);
-            }
-            else {
+            } else {
                 tmp->right += *tmp_int - prev;
                 i = g_list_next(i);
             }
@@ -183,19 +215,43 @@ static GPtrArray *array_time_slots(interval *I, GList *pairs){
     return array;
 }
 
+void create_ordered_jobs_array(GPtrArray *a, GPtrArray *b) {
+    interval *         tmp_interval;
+    Job *              tmp_j;
+    GPtrArray *        jobarray;
+    job_interval_pair *tmp_pair;
+    for (unsigned i = 0; i < a->len; ++i) {
+        tmp_interval = (interval *)g_ptr_array_index(a, i);
+        jobarray = tmp_interval->sigma;
+        for (unsigned j = 0; j < jobarray->len; ++j) {
+            tmp_j = (Job *)g_ptr_array_index(jobarray, j);
+            if (tmp_j->processingime <= tmp_interval->b) {
+                tmp_pair = CC_SAFE_MALLOC(1, job_interval_pair);
+                tmp_pair->j = tmp_j;
+                tmp_pair->I = tmp_interval;
+                tmp_pair->take = 0;
+                g_ptr_array_add(b, tmp_pair);
+            }
+        }
+    }
+
+    printf("There are %u layers\n", b->len);
+}
 
 int find_division(wctproblem *problem) {
-    int        val = 0;
-    int        njobs = problem->njobs;
-    int        tmp;
-    int        prev;
-    GPtrArray *tmp_array = g_ptr_array_new_with_free_func(g_interval_free);
-    GPtrArray *jobarray = problem->g_job_array;
-    Job *      tmp_j;
-    Job *      j1, *j2;
-    interval * tmp_interval;
+    int            val = 0;
+    int            counter = 0;
+    int            njobs = problem->njobs;
+    int            tmp;
+    int            prev;
+    wctdata *      root_pd = &(problem->root_pd);
+    GPtrArray *    tmp_array = g_ptr_array_new_with_free_func(g_interval_free);
+    GPtrArray *    jobarray = problem->g_job_array;
+    Job *          tmp_j;
+    Job *          j1, *j2;
+    interval *     tmp_interval;
     interval_pair *pair;
-    interval_pair tmp_pair;
+    interval_pair  tmp_pair;
 
     /** Find initial partition */
     prev = 0;
@@ -203,30 +259,32 @@ int find_division(wctproblem *problem) {
         tmp_j = (Job *)g_ptr_array_index(jobarray, i);
         tmp = CC_MIN(problem->H_max, tmp_j->duetime);
         if (prev < tmp) {
-            tmp_interval = interval_alloc(prev, tmp, jobarray, njobs);
+            tmp_interval = interval_alloc(prev, tmp, -1, jobarray, njobs);
             g_ptr_array_add(tmp_array, tmp_interval);
-            CCcheck_NULL_2(tmp_interval, "Failed to allocate memory") prev =
-                tmp_j->duetime;
+            CCcheck_NULL_2(tmp_interval, "Failed to allocate memory");
+            prev = tmp_j->duetime;
         }
     }
 
     if (prev < problem->H_max) {
-        tmp_interval = interval_alloc(prev, problem->H_max, jobarray, njobs);
+        tmp_interval =
+            interval_alloc(prev, problem->H_max, -1, jobarray, njobs);
         g_ptr_array_add(tmp_array, tmp_interval);
     }
 
     /** calculate the new intervals */
     for (unsigned i = 0; i < tmp_array->len; ++i) {
-        GList *pairs = (GList *) NULL;
-        tmp_interval = (interval *) g_ptr_array_index(tmp_array, i);
+        GList *pairs = (GList *)NULL;
+        tmp_interval = (interval *)g_ptr_array_index(tmp_array, i);
         for (size_t j = 0; j < tmp_interval->sigma->len - 1; j++) {
             for (size_t k = j + 1; k < tmp_interval->sigma->len; k++) {
                 j1 = (Job *)g_ptr_array_index(tmp_interval->sigma, j);
                 j2 = (Job *)g_ptr_array_index(tmp_interval->sigma, k);
-                tmp_pair = (interval_pair) {j1, j2};
+                tmp_pair = (interval_pair){j1, j2};
                 if (!check_interval(&tmp_pair, i, tmp_array)) {
                     pair = CC_SAFE_MALLOC(1, interval_pair);
-                    *pair =  (interval_pair) {j1, j2, tmp_pair.left, tmp_pair.right};
+                    *pair =
+                        (interval_pair){j1, j2, tmp_pair.left, tmp_pair.right};
                     pairs = g_list_append(pairs, pair);
                 }
             }
@@ -235,16 +293,21 @@ int find_division(wctproblem *problem) {
         if (pairs) {
             GPtrArray *slots;
             slots = array_time_slots(tmp_interval, pairs);
-            for(unsigned j = 1; j < slots->len; ++j) {
-                g_ptr_array_add(problem->e, interval_alloc(*((int *)slots->pdata[j - 1]), *((int *)slots->pdata[j]), jobarray, njobs));
+            for (unsigned j = 1; j < slots->len; ++j) {
+                g_ptr_array_add(root_pd->local_intervals,
+                                interval_alloc(*((int *)slots->pdata[j - 1]),
+                                               *((int *)slots->pdata[j]),
+                                               counter, jobarray, njobs));
+                counter++;
             }
             g_ptr_array_free(slots, TRUE);
         } else {
-            g_ptr_array_add(problem->e, interval_alloc(tmp_interval->a, tmp_interval->b, jobarray, njobs));
+            g_ptr_array_add(root_pd->local_intervals,
+                            interval_alloc(tmp_interval->a, tmp_interval->b,
+                                           counter, jobarray, njobs));
+            counter++;
         }
     }
-
-    g_ptr_array_foreach(problem->e, g_print_interval, NULL);
 
 CLEAN:
     g_ptr_array_free(tmp_array, TRUE);
