@@ -1,47 +1,26 @@
-#include <wct.h>
+#include <wctprivate.h>
 #include <PricerSolver.hpp>
 #include <iostream>
 #include <vector>
 
 template <typename T = double, bool reverse = false>
-int construct_sol(scheduleset **       set,
-                  int *                nnewsets,
-                  Job *                jobarray,
-                  Optimal_Solution<T> &sol,
-                  int                  nbjobs) {
-    int               val = 0;
-    int               nbset = 1;
-    int               C = 0;
-    std::vector<int> *v = &(sol.jobs);
-    scheduleset *     newset = CC_SAFE_MALLOC(1, scheduleset);
-    CCcheck_NULL_2(newset, "Failed to allocate memory newset");
-    scheduleset_init(newset);
-    newset->members = CC_SAFE_MALLOC(sol.jobs.size() + 1, int);
-    CCcheck_NULL_2(newset->members, "Failed to allocate memory members");
-    newset->C = CC_SAFE_MALLOC(sol.jobs.size(), int);
-    CCcheck_NULL_2(newset->C, "Failed to allocate memory");
-    newset->table = g_hash_table_new(g_direct_hash, g_direct_equal);
-    CCcheck_NULL_2(newset->table, "Failed to allocate memory");
+int construct_sol(wctdata *pd, Optimal_Solution<T> &sol) {
+    int          val = 0;
+    int          nbset = 1;
+    Job *        tmp_j;
+    scheduleset *newset = scheduleset_alloc(pd->njobs);
+    CCcheck_NULL_3(newset, "Failed to allocate memory newset");
 
-    if (reverse) {
-        std::copy(v->rbegin(), v->rend(), newset->members);
-    } else {
-        std::copy(v->begin(), v->end(), newset->members);
-    }
-
-    for (size_t i = 0; i < sol.jobs.size(); i++) {
-        C += jobarray[newset->members[i]].processingime;
-        newset->C[i] = C;
-        g_hash_table_insert(newset->table, GINT_TO_POINTER(newset->members[i]),
-                            newset->C + i);
+    for (unsigned i = 0; i < sol.jobs->len; ++i) {
+        tmp_j = (Job *)g_ptr_array_index(sol.jobs, i);
+        g_ptr_array_add(newset->jobs, tmp_j);
+        g_hash_table_add(newset->table, tmp_j);
     }
 
     newset->totwct = sol.cost;
     newset->totweight = sol.C_max;
-    newset->count = sol.jobs.size();
-    newset->members[sol.jobs.size()] = nbjobs;
-    *set = newset;
-    *nnewsets = 1;
+    pd->newsets = newset;
+    pd->nnewsets = 1;
 CLEAN:
 
     if (val) {
@@ -52,12 +31,14 @@ CLEAN:
 }
 
 extern "C" {
-PricerSolver *newSolver(Job *jobarray, int nbjobs, int Hmin, int Hmax) {
-    return new PricerSolver(jobarray, nbjobs, Hmin, Hmax);
-}
 
-PricerSolver *newSolverDP(Job *jobarray, int nbjobs, int Hmin, int Hmax) {
-    return new PricerSolver(jobarray, nbjobs, Hmin, Hmax, false);
+static double compute_lagrange(Optimal_Solution<double> &sol,
+                               double *                  rhs,
+                               double *                  pi,
+                               int                       nbjobs);
+
+PricerSolver *newSolver(GPtrArray *jobs, GPtrArray *ordered_jobs, int nmachines, int ub) {
+    return new PricerSolver(jobs, ordered_jobs, nmachines, ub);
 }
 
 PricerSolver *copySolver(PricerSolver *src) { return new PricerSolver(*src); }
@@ -68,29 +49,13 @@ void print_dot_file(PricerSolver *solver, char *name) {
 
 void freeSolver(PricerSolver *src) { delete src; }
 
-int solve_dynamic_programming_ahv(wctdata *pd) {
-    int                      val = 0;
-    Optimal_Solution<double> s = pd->solver->dynamic_programming_ahv(pd->pi);
-
-    if (s.obj < -0.00001) {
-        val = construct_sol<double, true>(&(pd->newsets), &(pd->nnewsets),
-                                          pd->jobarray, s, pd->njobs);
-        CCcheck_val_2(val, "Failed in constructing sol");
-    } else {
-        pd->nnewsets = 0;
-    }
-
-CLEAN:
-    return val;
-}
-
 int solve_farkas_dbl(wctdata *pd) {
     int                      val = 0;
     Optimal_Solution<double> s = pd->solver->solve_farkas_double(pd->pi);
 
     if (s.obj < -0.00001) {
-        val = construct_sol(&(pd->newsets), &(pd->nnewsets), pd->jobarray, s,
-                            pd->njobs);
+        // val = construct_sol(&(pd->newsets), &(pd->nnewsets), pd->jobarray, s,
+        //                     pd->njobs);
         CCcheck_val_2(val, "Failed in constructing jobs");
     } else {
         pd->nnewsets = 0;
@@ -100,53 +65,74 @@ CLEAN:
     return val;
 }
 
-int solve_farkas_dbl_DP(wctdata *pd) {
-    int                      val = 0;
-    Optimal_Solution<double> s =
-        pd->solver->dynamic_programming_ahv_farkas(pd->pi);
+static double compute_lagrange(Optimal_Solution<double> &sol,
+                               double *                  rhs,
+                               double *                  pi,
+                               int                       nbjobs) {
+    double result = 0.0;
+    double a = 0.0;
+    Job *  tmp_j;
 
-    if (s.obj < -0.00001) {
-        val = construct_sol(&(pd->newsets), &(pd->nnewsets), pd->jobarray, s,
-                            pd->njobs);
-        CCcheck_val_2(val, "Failed in constructing jobs");
-    } else {
-        pd->nnewsets = 0;
+    for (unsigned i = 0; i < sol.jobs->len; ++i) {
+        tmp_j = (Job *)g_ptr_array_index(sol.jobs, i);
+        result += pi[tmp_j->job];
     }
 
-CLEAN:
+    for (int i = 0; i < nbjobs; ++i) {
+        a += rhs[i] * pi[i];
+    }
+
+    result = CC_MIN(0.0, sol.cost - result);
+    result = rhs[nbjobs] * result;
+    result += a;
+    return result;
+}
+
+static double compute_reduced_cost(Optimal_Solution<double> &sol,
+                               double *                  pi,
+                               int                       nbjobs) {
+    double result = pi[nbjobs];
+    Job *  tmp_j;
+    int C = 0;
+
+    for (unsigned i = 0; i < sol.jobs->len; ++i) {
+        tmp_j = (Job *)g_ptr_array_index(sol.jobs, i);
+        C += tmp_j->processingime;
+        result += pi[tmp_j->job] - value_Fj(C, tmp_j);
+    }
+
+    return result;
+}
+
+int evaluate_nodes(wctdata *pd) {
+    int    val = 0;
+    int    UB = pd->problem->opt_sol->tw;
+    double LB = pd->LP_lower_bound;
+    int    nmachines = pd->problem->nmachines;
+
+    pd->solver->evaluate_nodes(pd->pi, UB, LB, nmachines, pd->reduced_cost);
+
     return val;
 }
 
-int solve_weight_dbl_bdd(wctdata *pd) {
-    int                      val = 0;
-    Optimal_Solution<double> s = pd->solver->solve_weight_bdd_double(pd->pi);
+int calculate_new_ordered_jobs(wctdata *pd){
+    int val = 0;
 
-    if (s.obj > 0.00001) {
-        val = construct_sol(&(pd->newsets), &(pd->nnewsets), pd->jobarray, s,
-                            pd->njobs);
-        CCcheck_val_2(val, "Failed in construction")
-    } else {
-        pd->nnewsets = 0;
-    }
+    pd->solver->calculate_new_ordered_jobs();
 
-CLEAN:
     return val;
 }
 
-int solve_weight_dbl_zdd(wctdata *pd) {
-    int                      val = 0;
-    Optimal_Solution<double> s = pd->solver->solve_weight_zdd_double(pd->pi);
+int build_solve_mip(wctdata *pd) {
+    int val = 0;
 
-    if (s.obj > 0.00001) {
-        val = construct_sol(&(pd->newsets), &(pd->nnewsets), pd->jobarray, s,
-                            pd->njobs);
-        CCcheck_val_2(val, "Failed in construction")
-    } else {
-        pd->nnewsets = 0;
-    }
+    pd->solver->build_mip();
 
-CLEAN:
     return val;
+}
+
+void print_number_nodes_edges(wctdata *pd){
+    pd->solver->print_number_nodes_edges();
 }
 
 void deletePricerSolver(PricerSolver *solver) {
@@ -158,27 +144,8 @@ void deletePricerSolver(PricerSolver *solver) {
 int calculate_table(PricerSolver *solver, wctparms *parms) {
     int val = 0;
 
-    switch (parms->solver) {
-        case bdd_solver:
-            solver->init_bdd_table();
-            break;
-
-        case zdd_solver:
-            solver->init_zdd_table();
-            break;
-
-        case DP_solver:
-            break;
-    }
-
-    switch (parms->construct) {
-        case yes_construct:
-            break;
-
-        case no_construct:
-            solver->init_table_farkas();
-            break;
-    }
+    solver->init_zdd_table();
+    solver->init_table_farkas();
 
     return val = 0;
 }
@@ -191,27 +158,15 @@ int add_conflict_constraints(PricerSolver *solver,
                              int           ecount_differ) {
     int val = 0;
 
-    switch (parms->solver) {
-        case bdd_solver:
-            solver->init_bdd_conflict_solver(elist_same, ecount_same,
-                                             elist_differ, ecount_differ);
-            break;
-
-        case zdd_solver:
-            solver->init_zdd_conflict_solver(elist_same, ecount_same,
-                                             elist_differ, ecount_differ);
-            break;
-
-        case DP_solver:
-            break;
-    }
+    solver->init_zdd_conflict_solver(elist_same, ecount_same, elist_differ,
+                                     ecount_differ);
 
     return val;
 }
 
-void iterate_dd(PricerSolver *solver) { solver->iterate_dd(); }
-
 void iterate_zdd(PricerSolver *solver) { solver->iterate_zdd(); }
+
+void print_number_paths(PricerSolver *solver){ solver->print_number_paths(); }
 
 int free_conflict_constraints(PricerSolver *solver,
                               wctparms *    parms,
@@ -219,48 +174,22 @@ int free_conflict_constraints(PricerSolver *solver,
                               int           ecount_differ) {
     int val = 0;
 
-    switch (parms->solver) {
-        case bdd_solver:
-            solver->free_bdd_solver(ecount_same, ecount_differ);
-            break;
-
-        case zdd_solver:
-            solver->free_zdd_solver(ecount_same, ecount_differ);
-            break;
-
-        case DP_solver:
-            break;
-    }
+    solver->free_zdd_solver(ecount_same, ecount_differ);
 
     return val;
 }
 
-size_t get_datasize(PricerSolver *solver) { return solver->zdd->size(); }
+size_t get_datasize(PricerSolver *solver) { return solver->get_datasize(); }
 
 size_t get_numberrows_zdd(PricerSolver *solver) {
-    return solver->zdd->root().row();
-}
-
-size_t get_numberrows_bdd(PricerSolver *solver) {
-    return solver->dd->root().row();
+    return solver->get_numberrows_zdd();
 }
 
 int add_one_conflict(
-    PricerSolver *solver, wctparms *parms, int v1, int v2, int same) {
+    PricerSolver *solver, wctparms *parms, Job *v1, Job *v2, int same) {
     int val = 0;
 
-    switch (parms->solver) {
-        case bdd_solver:
-            solver->init_bdd_one_conflict(v1, v2, same);
-            break;
-
-        case zdd_solver:
-            solver->init_zdd_one_conflict(v1, v2, same);
-            break;
-
-        case DP_solver:
-            break;
-    }
+    solver->init_zdd_one_conflict(v1->job, v2->job, same);
 
     return val;
 }
@@ -271,10 +200,6 @@ int init_tables(PricerSolver *solver) {
     return val;
 }
 
-void set_release_due_time(PricerSolver *solver, Job *jobarray) {
-    solver->set_release_due_time(jobarray);
-}
-
 static void compute_subgradient(Optimal_Solution<double> &sol,
                                 double *                  sub_gradient,
                                 double *                  rhs,
@@ -283,9 +208,9 @@ static void compute_subgradient(Optimal_Solution<double> &sol,
     fill_dbl(sub_gradient, nbjobs, 1.0);
     sub_gradient[nbjobs] = nbmachines;
 
-    for (auto &v : sol.jobs) {
-        sub_gradient[v] -= (double)nbmachines * 1.0;
-    }
+    // for (auto &v : sol.jobs) {
+    //     sub_gradient[v] -= (double)nbmachines * 1.0;
+    // }
 }
 
 static void adjust_alpha(double *pi_out,
@@ -324,167 +249,92 @@ static void compute_pi_eta_sep(int     vcount,
     *eta_sep = alpha * (*eta_in) + beta * (*eta_out);
 }
 
-int solve_pricing(wctdata *pd, wctparms *parms) {
+int solve_pricing(wctdata *pd, wctparms *parms, int evaluate) {
     int val = 0;
 
-    switch (parms->solver) {
-        case bdd_solver:
-            val = solve_weight_dbl_bdd(pd);
-            CCcheck_val_2(val, "Failed solve_weight_dbl_bdd");
-            break;
 
-        case zdd_solver:
-            val = solve_weight_dbl_zdd(pd);
-            CCcheck_val_2(val, "Failed solve_weight_dbl_zdd") break;
+    Optimal_Solution<double> sol = pd->solver->solve_weight_zdd_double(pd->pi);
+    if(sol.obj > 0.000001) {
+       val = construct_sol(pd, sol);
+       CCcheck_val_2(val, "Failed in construction")
+    } else {
+        pd->nnewsets = 0;
+    }
 
-        case DP_solver:
-            val = solve_dynamic_programming_ahv(pd);
-            CCcheck_val_2(val, "Failed in solve_dynamic_programming_ahv") break;
+    if(pd->iterations%5 == 0 || evaluate) {
+        pd->reduced_cost = compute_reduced_cost(sol, pd->pi, pd->njobs);
+        pd->solver->evaluate_nodes(pd->pi, pd->problem->opt_sol->tw, pd->LP_lower_bound_BB , pd->problem->nmachines, pd->reduced_cost);
+    }
+
+    if(pd->iterations%pd->njobs == 0) {
+        print_number_nodes_edges(pd);
     }
 
 CLEAN:
     return val;
 }
 
-static double compute_lagrange(Optimal_Solution<double> &sol,
-                               double *                  rhs,
-                               double *                  pi,
-                               int                       nbjobs,
-                               int                       nbmachines) {
-    double            result = 0;
-    double            a = 0.0;
-    int               i;
-    std::vector<int> *v = &(sol.jobs);
-
-    for (std::vector<int>::iterator it = v->begin(); it != v->end(); ++it) {
-        result += pi[*it];
-    }
-
-    for (i = 0; i < nbjobs; ++i) {
-        a += rhs[i] * pi[i];
-    }
-
-    // a += rhs[nbjobs] * pi[nbjobs];
-    result = CC_MIN(0.0, sol.cost - result);
-    result = rhs[nbjobs] * result;
-    result += a;
-    return result;
-}
-
-static void compute_sol_stab(PricerSolver *            solver,
-                             wctparms *                parms,
-                             double *                  pi,
-                             Optimal_Solution<double> *sol) {
-    switch (parms->solver) {
-        case bdd_solver:
-            *sol = solver->solve_weight_bdd_double(pi);
-            break;
-
-        case zdd_solver:
-            *sol = solver->solve_weight_zdd_double(pi);
-            break;
-
-        case DP_solver:
-            *sol = solver->dynamic_programming_ahv(pi);
-            break;
-    }
-}
-
-static int construct_sol_stab(wctdata *                 pd,
-                              wctparms *                parms,
-                              Optimal_Solution<double> &sol) {
-    int val = 0;
-
-    switch (parms->solver) {
-        case bdd_solver:
-        case zdd_solver:
-            val = construct_sol(&(pd->newsets), &(pd->nnewsets), pd->jobarray,
-                                sol, pd->njobs);
-            CCcheck_val_2(val, "Failed in construction of solution");
-            break;
-
-        case DP_solver:
-            val = construct_sol<double, true>(&(pd->newsets), &(pd->nnewsets),
-                                              pd->jobarray, sol, pd->njobs);
-            CCcheck_val_2(val, "Failed in construction of solution");
-            break;
-    }
-
-CLEAN:
-    return val;
-}
 
 int solve_stab(wctdata *pd, wctparms *parms) {
-    int                      val = 0;
-    int                      heading_in = 0;
-    PricerSolver *           solver = pd->solver;
-    Optimal_Solution<double> sol;
-    heading_in =
-        (pd->eta_in == 0.0)
-            ? 1
-            : !(CC_OURABS((pd->eta_out - pd->eta_in) / (pd->eta_in)) < 4.0);
-
-    if (heading_in) {
-        double result;
-        compute_sol_stab(solver, parms, pd->pi, &sol);
-        result =
-            compute_lagrange(sol, pd->rhs, pd->pi, pd->njobs, pd->nmachines);
-
-        if (result > pd->eta_in) {
-            pd->eta_in = result;
-            memcpy(pd->pi_in, pd->pi, sizeof(double) * (pd->njobs + 1));
-        }
-
-        val = construct_sol_stab(pd, parms, sol);
-        CCcheck_val_2(val, "Failed in construct solution");
-    } else {
+    int           val = 0;
+    PricerSolver *solver = pd->solver;
         double k = 0.0;
         double alpha;
         bool   mispricing = false;
         double result_sep;
-        double result_out;
 
+    if((pd->LP_lower_bound_BB - pd->eta_in)/(pd->LP_lower_bound_BB) > 0.5) {
+        Optimal_Solution<double> sol = solver->solve_weight_zdd_double(pd->pi_out);
+        result_sep = compute_lagrange(sol, pd->rhs, pd->pi_out, pd->njobs);
+        pd->reduced_cost = compute_reduced_cost(sol, pd->pi, pd->njobs);
+        if(pd->reduced_cost >= 0.00001) {
+            val = construct_sol(pd, sol);
+            CCcheck_val_2(val, "Failed in construct_sol_stab");
+            pd->update = 1;
+        }
+    } else {
         do {
             k += 1.0;
             alpha = CC_MAX(0, 1.0 - k * (1.0 - pd->alpha));
             compute_pi_eta_sep(pd->njobs, pd->pi_sep, &(pd->eta_sep), alpha,
                                pd->pi_in, &(pd->eta_in), pd->pi_out,
                                &(pd->eta_out));
-            compute_sol_stab(solver, parms, pd->pi_sep, &sol);
-            result_sep = compute_lagrange(sol, pd->rhs, pd->pi_sep, pd->njobs,
-                                          pd->nmachines);
+            Optimal_Solution<double> sol = solver->solve_weight_zdd_double(pd->pi_sep);
+            result_sep = compute_lagrange(sol, pd->rhs, pd->pi_sep, pd->njobs);
+            pd->reduced_cost = compute_reduced_cost(sol, pd->pi_out, pd->njobs);
 
-            if (result_sep < pd->eta_sep) {
-                val = construct_sol_stab(pd, parms, sol);
+            if (pd->reduced_cost >= 0.00001) {
+                val = construct_sol(pd, sol);
                 CCcheck_val_2(val, "Failed in construct_sol_stab");
                 pd->update = 1;
                 mispricing = false;
             } else {
-                result_out = compute_lagrange(sol, pd->rhs, pd->pi_out,
-                                              pd->njobs, pd->nmachines);
+                pd->reduced_cost =compute_reduced_cost(sol, pd->pi_out, pd->njobs);
 
-                if (result_out < pd->eta_out) {
-                    val = construct_sol_stab(pd, parms, sol);
+                if (pd->reduced_cost < 0.000001) {
                     CCcheck_val_2(val, "Failed in construct_sol_stab");
-                    mispricing = false;
+                    mispricing = true;
                     pd->update = 0;
                 } else {
-                    mispricing = true;
+                    val = construct_sol(pd, sol);
+                    mispricing = false;
+                    pd->update = 1;
                 }
             }
         } while (mispricing && alpha > 0); /** mispricing check */
-
-        if (result_sep > pd->eta_in) {
-            pd->eta_in = result_sep;
-            memcpy(pd->pi_in, pd->pi_sep, sizeof(double) * (pd->njobs + 1));
-        }
     }
 
-    if (dbg_lvl() > 0) {
+
+    if (result_sep > pd->eta_in) {
+        pd->eta_in = result_sep;
+        memcpy(pd->pi_in, pd->pi_sep, sizeof(double) * (pd->njobs + 1));
+    }
+
+    if (pd->iterations%pd->njobs == 0) {
         printf(
-            "heading = %d, alpha = %f, result of primal bound and Lagragian "
+            "alpha = %f, result of primal bound and Lagragian "
             "bound: out =%f, in = %f\n",
-            heading_in, pd->alpha, pd->eta_out, pd->eta_in);
+             pd->alpha, pd->eta_out, pd->eta_in);
     }
 
 CLEAN:
@@ -492,14 +342,13 @@ CLEAN:
 }
 
 int solve_stab_dynamic(wctdata *pd, wctparms *parms) {
-    int                      val = 0;
-    PricerSolver *           solver = pd->solver;
-    Optimal_Solution<double> sol;
-    double                   k = 0.0;
-    double                   alpha;
-    double                   result_sep;
-    double                   result_out;
-    bool                     mispricing = false;
+    int           val = 0;
+    PricerSolver *solver = pd->solver;
+    double        k = 0.0;
+    double        alpha;
+    double        result_sep;
+    double        result_out;
+    bool          mispricing = false;
 
     do {
         k += 1.0;
@@ -507,26 +356,25 @@ int solve_stab_dynamic(wctdata *pd, wctparms *parms) {
         compute_pi_eta_sep(pd->njobs, pd->pi_sep, &(pd->eta_sep), alpha,
                            pd->pi_in, &(pd->eta_in), pd->pi_out,
                            &(pd->eta_out));
-        compute_sol_stab(solver, parms, pd->pi_sep, &(sol));
-        result_sep = compute_lagrange(sol, pd->rhs, pd->pi_sep, pd->njobs,
-                                      pd->nmachines);
+        Optimal_Solution<double> sol =
+            solver->solve_weight_zdd_double(pd->pi_sep);
+        result_sep = compute_lagrange(sol, pd->rhs, pd->pi_sep, pd->njobs);
 
         if (result_sep < pd->eta_sep) {
-            val = construct_sol_stab(pd, parms, sol);
+            val = construct_sol(pd, sol);
             CCcheck_val_2(val, "Failed in construct_sol_stab");
-            pd->update = 1;
             compute_subgradient(sol, pd->subgradient, pd->rhs, pd->njobs,
                                 pd->nmachines);
             adjust_alpha(pd->pi_out, pd->pi_in, pd->subgradient, pd->njobs,
                          alpha);
             pd->alpha = alpha;
+            pd->update = 1;
             mispricing = false;
         } else {
-            result_out = compute_lagrange(sol, pd->rhs, pd->pi_out, pd->njobs,
-                                          pd->nmachines);
+            result_out = compute_lagrange(sol, pd->rhs, pd->pi_out, pd->njobs);
 
             if (result_out < pd->eta_out) {
-                val = construct_sol_stab(pd, parms, sol);
+                val = construct_sol(pd, sol);
                 CCcheck_val_2(val, "Failed in construct_sol_stab");
                 mispricing = false;
                 pd->update = 0;
